rbac:extract echo configuration from main and move to new pkg/router location
