+++ cmd/content-sources/main.go
@@ -10,7 +10,6 @@ import (
 	"syscall"
 	"time"
 
-	"github.com/content-services/content-sources-backend/pkg/client"
 	"github.com/content-services/content-sources-backend/pkg/config"
 	"github.com/content-services/content-sources-backend/pkg/db"
 	"github.com/content-services/content-sources-backend/pkg/event"
@@ -18,15 +17,13 @@ import (
 	"github.com/content-services/content-sources-backend/pkg/handler"
 	m "github.com/content-services/content-sources-backend/pkg/instrumentation"
 	custom_collector "github.com/content-services/content-sources-backend/pkg/instrumentation/custom"
-	"github.com/content-services/content-sources-backend/pkg/middleware"
+	"github.com/content-services/content-sources-backend/pkg/router"
 	mocks_rbac "github.com/content-services/content-sources-backend/pkg/test/mocks/rbac"
 	"github.com/labstack/echo/v4"
 	echo_middleware "github.com/labstack/echo/v4/middleware"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
-	"github.com/redhatinsights/platform-go-middlewares/identity"
 	"github.com/rs/zerolog/log"
-	"github.com/ziflex/lecho/v3"
 )
 
 func main() {
@@ -96,7 +93,7 @@ func kafkaConsumer(ctx context.Context, wg *sync.WaitGroup, metrics *m.Metrics) 
 func apiServer(ctx context.Context, wg *sync.WaitGroup, allRoutes bool, metrics *m.Metrics) {
 	wg.Add(2) // api server & shutdown monitor
 
-	echo := config.ConfigureEchoWithMetrics(metrics)
+	echo := router.ConfigureEchoWithMetrics(metrics)
 	handler.RegisterPing(echo)
 	if allRoutes {
 		handler.RegisterRoutes(echo)
@@ -123,7 +120,7 @@ func apiServer(ctx context.Context, wg *sync.WaitGroup, allRoutes bool, metrics 
 
 func instrumentation(ctx context.Context, wg *sync.WaitGroup, metrics *m.Metrics) {
 	wg.Add(2)
-	e := ConfigureEcho(false)
+	e := router.ConfigureEcho(false)
 
 	metricsPath := config.Get().Metrics.Path
 	metricsPort := config.Get().Metrics.Port
@@ -170,50 +167,6 @@ func instrumentation(ctx context.Context, wg *sync.WaitGroup, metrics *m.Metrics
 		<-custom_ctx.Done()
 		custom_cancel()
 	}()
-}
-
-func ConfigureEcho(allRoutes bool) *echo.Echo {
-	e := echo.New()
-	echoLogger := lecho.From(log.Logger,
-		lecho.WithTimestamp(),
-		lecho.WithCaller(),
-	)
-	e.Use(echo_middleware.AddTrailingSlash(), echo_middleware.Logger())
-	e.Use(echo_middleware.RequestIDWithConfig(echo_middleware.RequestIDConfig{
-		TargetHeader: "x-rh-insights-request-id",
-	}))
-	e.Use(lecho.Middleware(lecho.Config{
-		Logger:       echoLogger,
-		RequestIDKey: "x-rh-insights-request-id",
-	}))
-
-	// Liveness and readiness are set up before security checks
-	handler.RegisterPing(e)
-
-	e.Use(middleware.WrapMiddlewareWithSkipper(identity.EnforceIdentity, middleware.SkipLiveness))
-	if config.Get().Clients.RbacEnabled {
-		rbacBaseUrl := config.Get().Clients.RbacBaseUrl
-		rbacTimeout := time.Duration(int64(config.Get().Clients.RbacTimeout) * int64(time.Second))
-		rbacClient := client.NewRbac(rbacBaseUrl, rbacTimeout)
-		log.Info().Msgf("rbacBaseUrl=%s", rbacBaseUrl)
-		log.Info().Msgf("rbacTimeout=%d secs", rbacTimeout/time.Second)
-		e.Use(
-			middleware.NewRbac(
-				middleware.Rbac{
-					BaseUrl:        config.Get().Clients.RbacBaseUrl,
-					Skipper:        middleware.SkipLiveness,
-					PermissionsMap: middleware.ServicePermissions,
-					Client:         rbacClient,
-				},
-			),
-		)
-	}
-
-	if allRoutes {
-		handler.RegisterRoutes(e)
-	}
-	e.HTTPErrorHandler = config.CustomHTTPErrorHandler
-	return e
 }
 
 func mockRbac(ctx context.Context, wg *sync.WaitGroup) {
